back/requirements.txt
Django==3.1.4
django-ckeditor==6.0.0
Pillow==8.0.1
psycopg2-binary==2.8.6